完成三种角色列表查看以及基本信息查询功能

+++ src/db/models/t-user.js
@@ -17,10 +17,10 @@ export default db.define('t_user', {
   department: Sequelize.STRING(32), // 科室
   departmentUuid: Sequelize.STRING(36), // 科室uuid
   // 以下字段仅供普通员工使用
-  verifyStatus: Sequelize.STRING(32), // 审核状态
+  verifyStatus: Sequelize.STRING(32), // 核实状态(未填写完毕/未核实/核实不通过/核实通过)
   verifyTime: Sequelize.DATE, // 最新核实时间
   reviewTime: Sequelize.DATE, // 评审时间
-  totalScore: Sequelize.FLOAT(3, 2), // 总得分
+  totalScore: Sequelize.FLOAT(5, 2), // 总得分
   lastWriteTime: Sequelize.DATE, // 上次填写时间
   currentWriteTime: Sequelize.DATE, // 最新填写时间
 });
